+++ src/AbstractMoney.php
@@ -9,7 +9,7 @@ use Brick\Math\Exception\MathException;
 use Brick\Math\Exception\RoundingNecessaryException;
 use Brick\Math\RoundingMode;
 use Brick\Money\Exception\ContextException;
-use Brick\Money\Exception\MoneyMismatchException;
+use Brick\Money\Exception\CurrencyMismatchException;
 use JsonSerializable;
 use Override;
 use Stringable;
@@ -129,8 +129,8 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
      *
      * @return -1|0|1 If `$this` is less than, equal to, or greater than `$that`.
      *
-     * @throws MathException          If the argument is an invalid number.
-     * @throws MoneyMismatchException If the argument is a money in a different currency.
+     * @throws MathException             If the argument is an invalid number.
+     * @throws CurrencyMismatchException If the argument is a money in a different currency.
      *
      * @pure
      */
@@ -145,8 +145,8 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
      * This method throws an exception if the argument is a money in a different currency.
      * If you want to return false when the currencies differ, use isSameValueAs() instead.
      *
-     * @throws MathException          If the argument is an invalid number.
-     * @throws MoneyMismatchException If the argument is a money in a different currency.
+     * @throws MathException             If the argument is an invalid number.
+     * @throws CurrencyMismatchException If the argument is a money in a different currency.
      *
      * @pure
      */
@@ -158,8 +158,8 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
     /**
      * Returns whether this money is less than the given amount.
      *
-     * @throws MathException          If the argument is an invalid number.
-     * @throws MoneyMismatchException If the argument is a money in a different currency.
+     * @throws MathException             If the argument is an invalid number.
+     * @throws CurrencyMismatchException If the argument is a money in a different currency.
      *
      * @pure
      */
@@ -171,8 +171,8 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
     /**
      * Returns whether this money is less than or equal to the given amount.
      *
-     * @throws MathException          If the argument is an invalid number.
-     * @throws MoneyMismatchException If the argument is a money in a different currency.
+     * @throws MathException             If the argument is an invalid number.
+     * @throws CurrencyMismatchException If the argument is a money in a different currency.
      *
      * @pure
      */
@@ -184,8 +184,8 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
     /**
      * Returns whether this money is greater than the given amount.
      *
-     * @throws MathException          If the argument is an invalid number.
-     * @throws MoneyMismatchException If the argument is a money in a different currency.
+     * @throws MathException             If the argument is an invalid number.
+     * @throws CurrencyMismatchException If the argument is a money in a different currency.
      *
      * @pure
      */
@@ -197,8 +197,8 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
     /**
      * Returns whether this money is greater than or equal to the given amount.
      *
-     * @throws MathException          If the argument is an invalid number.
-     * @throws MoneyMismatchException If the argument is a money in a different currency.
+     * @throws MathException             If the argument is an invalid number.
+     * @throws CurrencyMismatchException If the argument is a money in a different currency.
      *
      * @pure
      */
@@ -211,7 +211,7 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
      * Returns whether this money's amount and currency are equal to those of the given money.
      *
      * Unlike isEqualTo(), this method only accepts a money, and returns false if the given money is in another
-     * currency, instead of throwing a MoneyMismatchException.
+     * currency, instead of throwing a CurrencyMismatchException.
      *
      * @pure
      */
@@ -263,7 +263,7 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
      *
      * @param AbstractMoney|BigNumber|int|string $that A money or amount.
      *
-     * @throws MoneyMismatchException If currencies don't match.
+     * @throws CurrencyMismatchException If currencies don't match.
      *
      * @pure
      */
@@ -271,7 +271,7 @@ abstract readonly class AbstractMoney implements Monetary, Stringable, JsonSeria
     {
         if ($that instanceof AbstractMoney) {
             if (! $that->getCurrency()->isEqualTo($this->getCurrency())) {
-                throw MoneyMismatchException::currencyMismatch($this->getCurrency(), $that->getCurrency());
+                throw CurrencyMismatchException::currencyMismatch($this->getCurrency(), $that->getCurrency());
             }
 
             return $that->getAmount();